requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
decorator==4.4.2
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.1.0
MarkupSafe==2.0.1
moviepy==1.0.3
nltk==3.6.3
numpy==1.21.2
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
Pillow==8.3.2
proglog==0.1.9
regex==2021.10.8
requests==2.26.0
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.5.0
